Refactor and Add Responsive Styles
