+++ src/components/homeoptions/UpgradeSiteOption.jsx
@@ -1,12 +1,12 @@
-import Background from './shared/background/Background';
+import Background from '../shared/background/Background';
 
-import ShoppingSVG from '../images/person-studying-online4.png';
-import Ghost4PNG from '../images/PointingGhost.png';
+import ShoppingSVG from '../../images/person-studying-online4.png';
+import Ghost4PNG from '../../images/PointingGhost.png';
 
-import ResponsiveIcon from '../images/responsive-icon.svg';
-import AccessibleIcon from '../images/accessible-icon.svg';
-import SearchableIcon from '../images/searchable-icon.svg';
-import DesignSketch from '../images/design-sketch.webp';
+import ResponsiveIcon from '../../images/responsive-icon.svg';
+import AccessibleIcon from '../../images/accessible-icon.svg';
+import SearchableIcon from '../../images/searchable-icon.svg';
+import DesignSketch from '../../images/design-sketch.webp';
 
 function UpgradeSiteOption() {
     return (
