+++ src/components/homeoptions/_newsiteoption.scss
@@ -1,0 +1,71 @@
+.openingcall {
+    /* Layout */
+    width: 65vw;
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+    margin: 0 auto;
+
+    h2 {
+        /* Layout */
+        margin-bottom: 1rem;
+
+        /* Typography */
+        color: white;
+        font-family: 'Fredoka', sans-serif;
+        font-size: 4vw;
+        line-height: 1;
+    }
+}
+.blurb {
+    /* Layout */
+    display: grid;
+    justify-content: center;
+    align-items: center;
+    width: 90%;
+
+    /* Typography */
+    color: white;
+    font-size: 1.4vw;
+    font-family: 'Fredoka', sans-serif;
+    line-height: 1.6;
+
+    /* Design */
+    background-color: black;
+
+    p {
+        /* Typography */
+        font-weight: 100;
+        line-height: 2;
+
+        &::first-letter {
+            /* Large First Letter */
+            font-size: 4rem;
+
+            /* Fix Spacing from Larger Font */
+            line-height: 0;
+            letter-spacing: -0.75rem;
+            margin-left: -0.8rem;
+        }
+    }
+}
+
+.blurb-grid {
+    /* Layout */
+    grid-template-columns: 1fr 1.5fr;
+    grid-gap: 2rem;
+
+    /* Typography */
+    line-height: 2;
+}
+.designer-graphic {
+    display: flex;
+    align-items: center;
+    justify-content: center;
+
+    img {
+        aspect-ratio: 17 12;
+        width: 100%;
+    }
+}
